copy model definition enums from the generated code to avoid importing ffi; formatting
