+++ generator/integration-tests/basics/1.dart
@@ -4,7 +4,6 @@ import 'lib/objectbox.g.dart';
 import 'package:test/test.dart';
 import '../test_env.dart';
 import '../common.dart';
-import 'package:objectbox/src/native/bindings/bindings.dart';
 
 void main() {
   TestEnv<A> env;
