+++ generator/integration-tests/indexes/1.dart
@@ -6,7 +6,6 @@ import 'lib/objectbox.g.dart';
 import 'package:test/test.dart';
 import '../test_env.dart';
 import '../common.dart';
-import 'package:objectbox/src/native/bindings/bindings.dart';
 
 void main() {
   TestEnv<A> env;
