+++ generator/lib/src/code_chunks.dart
@@ -1,9 +1,6 @@
 import 'dart:convert';
 
-// TODO remove native imports
-
 import 'package:objectbox/src/modelinfo/index.dart';
-import 'package:objectbox/src/native/bindings/bindings.dart';
 import 'package:source_gen/source_gen.dart' show InvalidGenerationSourceError;
 
 class CodeChunks {
