+++ objectbox/lib/src/box.dart
@@ -1,2 +1,1 @@
-export 'native/box.dart'
-  if (dart.library.html) 'web/box.dart';
+export 'native/box.dart' if (dart.library.html) 'web/box.dart';
