+++ objectbox/lib/src/model.dart
@@ -1,2 +1,1 @@
-export 'native/model.dart'
-  if (dart.library.html) 'web/model.dart';
+export 'native/model.dart' if (dart.library.html) 'web/model.dart';
