+++ objectbox/lib/src/modelinfo/modelproperty.dart
@@ -1,7 +1,4 @@
-// TODO remove native imports
-
-import '../native/bindings/bindings.dart';
-import '../native/bindings/helpers.dart';
+import 'enums.dart';
 import 'iduid.dart';
 import 'modelentity.dart';
