+++ objectbox/lib/src/observable.dart
@@ -1,2 +1,1 @@
-export 'native/observable.dart'
-  if (dart.library.html) 'web/observable.dart';
+export 'native/observable.dart' if (dart.library.html) 'web/observable.dart';
