+++ objectbox/lib/src/sync.dart
@@ -1,2 +1,1 @@
-export 'native/sync.dart'
-  if (dart.library.html) 'web/sync.dart';
+export 'native/sync.dart' if (dart.library.html) 'web/sync.dart';
