+++ objectbox/lib/src/web/model.dart
@@ -1,0 +1,1 @@
+
