+++ objectbox/test/sync_test.dart
@@ -1,7 +1,7 @@
 import 'dart:math';
 import 'dart:typed_data';
 
-import 'package:objectbox/src/native/bindings/bindings.dart';
+import 'package:objectbox/src/modelinfo/enums.dart';
 import 'package:objectbox/objectbox.dart';
 import 'package:objectbox/internal.dart';
 import 'package:test/test.dart';
